Add CallCode byte code.
This byte code causes the virtual machine to push the next instruction's address to the top of the address stack, so that the VM will know what instruction to execute upon returning from the called function.  This byte code also sets the program counter to the first instruction of the called function.