Add the ReturnCode byte code.
The ReturnCode byte code returns program execution to the caller of the exiting function, by popping the return address from the caller off of the top of the address stack.  It also causes the virtual machine to pop the exiting function's activation frame, which is used by the debugger to display various information about the exiting function (if being run in debug mode).